% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPackagesLogs}
\alias{getPackagesLogs}
\title{Get package(s) logs configuration}
\usage{
getPackagesLogs()
}
\value{
data.frame.
}
\description{
Get package(s) logs configuration. Level and log file
}
\details{
Get package(s) logs configuration. Level and log file
}
\examples{
getPackagesLogs()
}

