% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setLayoutFormat}
\alias{setLayoutFormat}
\title{Set layout format}
\usage{
setLayoutFormat(layout.format = "[~time][~level] ~message")
}
\arguments{
\item{layout.format}{character. Layout format. Default to '[~time][~level] ~message'}
}
\value{
nothing. Assign value to LogEasieR.env
}
\description{
Set layout format
}
\details{
Available tags are :\cr
~level : Log level\cr
~time : Timestamp\cr
~father.env : Father Namespace\cr
~father.function : Father Calling function name\cr
~father.call : Full father call\cr
~children.env : Children Namespace\cr
~children.function : Calling function\cr
~children.call : Full chidren call\cr
~message : message\cr
}
\examples{
# default layout format want loading package
getLayoutFormat()

# change
setLayoutFormat(layout.format = '[~level][~time][~father.env-~father.function] ~message')
}

