% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setLogFile}
\alias{setLogFile}
\title{Set global log file}
\usage{
setLogFile(file = file.path(getwd(), "logeasier.log"))
}
\arguments{
\item{file}{character. path to log file. Default to file.path(getwd(), "logeasier.log")}
}
\value{
nothing. Assign value to LogEasieR.env
}
\description{
Set global log file
}
\details{
Set global log file
}
\examples{
# default log file want loading package
getLogFile()

# change
setLogFile(file = "mylogfile.txt")
}

