% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setLogLevel}
\alias{setLogLevel}
\title{Set global log level}
\usage{
setLogLevel(level = "INFO")
}
\arguments{
\item{level}{character. Value in c('INFO','WARN', 'STOP', 'NONE'). Default to "INFO"}
}
\value{
nothing. Assign value to LogEasieR.env
}
\description{
Set global log level
}
\details{
Set global log level
}
\examples{
# default level want loading package
getloglevel()

# change to WARN
setLogLevel(level = "WARN")
}

