% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setPackagesLogs}
\alias{setPackagesLogs}
\title{Set packages logs configuration}
\usage{
setPackagesLogs(package = NULL, level = getLogLevel()$global,
  logfile = getLogFile()$global)
}
\arguments{
\item{package}{character or character vector of package(s)}

\item{level}{character or character vector of levels for package(s). Value(s) in c('INFO','WARN', 'STOP', 'NONE'). Default to global log level}

\item{logfile}{character or character vector of log files for package(s). Default to global log file.}
}
\value{
nothing. Assign value to LogEasieR.env
}
\description{
Set packages logs configuration. Level and log file
}
\details{
Set packages logs configuration. Level and log file
}
\examples{
# disable logging for package FactoMineR
setPackagesLogs(package = "FactoMineR", level = "NONE")

# differents log files and levels for differents packages
setPackagesLogs(package = c("FactoMineR", "stats"), level = c("WARN","INFO"),
 logfile = c("factominer.log", "stats.txt"))
}

