% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setTimeStampFormat}
\alias{setTimeStampFormat}
\title{Set timestamp format}
\usage{
setTimeStampFormat(time.format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{time.format}{character. A time format. Default to ' \%Y-\%m-\%d \%H:\%M:\%S'. See \code{\link{strptime}}}
}
\value{
nothing. Assign value to LogEasieR.env
}
\description{
Set timestamp format
}
\details{
Set timestamp format
}
\examples{
# default timestamp format want loading package
getTimeStampFormat()

# change
setTimeStampFormat(time.format = "\%a \%b \%d \%X \%Y \%Z")
}

